Extract effect and sub-types into its own module. (#2266)
