+++ executor/src/witgen/jit/affine_symbolic_expression.rs
@@ -8,61 +8,12 @@ use itertools::Itertools;
 use num_traits::Zero;
 use powdr_number::FieldElement;
 
-use crate::witgen::EvalError;
+use crate::witgen::{jit::effect::Assertion, EvalError};
 
-use super::{super::range_constraints::RangeConstraint, symbolic_expression::SymbolicExpression};
-
-/// The effect of solving a symbolic equation.
-pub enum Effect<T: FieldElement, V> {
-    /// Variable can be assigned a value.
-    Assignment(V, SymbolicExpression<T, V>),
-    /// We learnt a new range constraint on variable.
-    RangeConstraint(V, RangeConstraint<T>),
-    /// A run-time assertion. If this fails, we have conflicting constraints.
-    Assertion(Assertion<T, V>),
-    /// a call to a different machine.
-    MachineCall(u64, Vec<MachineCallArgument<T, V>>),
-}
-
-/// A run-time assertion. If this fails, we have conflicting constraints.
-pub struct Assertion<T: FieldElement, V> {
-    pub lhs: SymbolicExpression<T, V>,
-    pub rhs: SymbolicExpression<T, V>,
-    /// If this is true, we assert that both sides are equal.
-    /// Otherwise, we assert that they are different.
-    pub expected_equal: bool,
-}
-
-impl<T: FieldElement, V> Assertion<T, V> {
-    pub fn assert_is_zero(condition: SymbolicExpression<T, V>) -> Effect<T, V> {
-        Self::assert_eq(condition, SymbolicExpression::from(T::from(0)))
-    }
-    pub fn assert_is_nonzero(condition: SymbolicExpression<T, V>) -> Effect<T, V> {
-        Self::assert_neq(condition, SymbolicExpression::from(T::from(0)))
-    }
-    pub fn assert_eq(lhs: SymbolicExpression<T, V>, rhs: SymbolicExpression<T, V>) -> Effect<T, V> {
-        Effect::Assertion(Assertion {
-            lhs,
-            rhs,
-            expected_equal: true,
-        })
-    }
-    pub fn assert_neq(
-        lhs: SymbolicExpression<T, V>,
-        rhs: SymbolicExpression<T, V>,
-    ) -> Effect<T, V> {
-        Effect::Assertion(Assertion {
-            lhs,
-            rhs,
-            expected_equal: false,
-        })
-    }
-}
-
-pub enum MachineCallArgument<T: FieldElement, V> {
-    Known(SymbolicExpression<T, V>),
-    Unknown(AffineSymbolicExpression<T, V>),
-}
+use super::{
+    super::range_constraints::RangeConstraint, effect::Effect,
+    symbolic_expression::SymbolicExpression,
+};
 
 #[derive(Default)]
 pub struct ProcessResult<T: FieldElement, V> {
