+++ executor/src/witgen/jit/block_machine_processor.rs
@@ -7,7 +7,7 @@ use powdr_number::FieldElement;
 use crate::witgen::{machines::MachineParts, FixedData};
 
 use super::{
-    affine_symbolic_expression::Effect,
+    effect::Effect,
     variable::Variable,
     witgen_inference::{FixedEvaluator, WitgenInference},
 };
@@ -155,7 +155,7 @@ mod test {
         constant_evaluator,
         witgen::{
             global_constraints,
-            jit::{affine_symbolic_expression::Effect, test_util::format_code},
+            jit::{effect::Effect, test_util::format_code},
             machines::{Connection, ConnectionKind, MachineParts},
             FixedData,
         },
