+++ executor/src/witgen/jit/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod affine_symbolic_expression;
 mod block_machine_processor;
 mod compiler;
+mod effect;
 pub(crate) mod function_cache;
 mod symbolic_expression;
 mod variable;
